% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_datasets.R
\docType{data}
\name{emw_12_clean}
\alias{emw_12_clean}
\title{Number of Children Data
A dataset that includes number of children for ever-married-women.}
\format{
A data frame with 26,616 rows and 60 columns.
\describe{
\item{IDind}{The IDind is a twelve-digit numeric variable that uniquely identified for all participants. Each participant will have the same ID in all datasets and in all survey years. }
\item{wave}{survey year}
\item{child_died}{Total number of Children who died}
\item{child_births}{Total number of Children given births to}
...
}
}
\source{
\url{https://www.cpc.unc.edu/projects/china}
}
\usage{
emw_12_clean
}
\description{
Number of Children Data
A dataset that includes number of children for ever-married-women.
}
\keyword{datasets}
