% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_datasets.R
\docType{data}
\name{indinc_10_clean}
\alias{indinc_10_clean}
\title{Annual Income from CHNS survey from 1989 to 2011}
\format{
\subsection{\code{indinc_10_clean}}{

A data frame with 62,029 rows and 32 columns:
\describe{
\item{IDind}{individual identification number}
\item{wave}{survey year}
\item{indwage}{annual wage, bonus, and other income}
}
}
}
\source{
\url{https://www.cpc.unc.edu/projects/china}
}
\usage{
indinc_10_clean
}
\description{
Annual Income from CHNS survey from 1989 to 2011
}
\keyword{datasets}
