% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_datasets.R
\docType{data}
\name{map_dat_clean}
\alias{map_dat_clean}
\title{Data set for mapping average income on China's map}
\format{
\subsection{\code{map_dat_clean}}{

A data frame that could map the average income by selected province and year:
\describe{
\item{province_en}{province names in english}
\item{wave}{survey year}
\item{wage_mean}{intial wage in unit of 1}
\item{name}{name of provinces in chinese}
\item{wage_mean_k}{wage in unit of 1000}
\item{wage_mean_cat}{mean wage in unit of 1000 divided into categories for mapping}
}
}
}
\source{
\url{https://www.cpc.unc.edu/projects/china}
}
\usage{
map_dat_clean
}
\description{
Data set for mapping average income on China's map
}
\keyword{datasets}
