% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_income.R
\name{real_income}
\alias{real_income}
\title{Function for deflating or inflating annual income to an inputted base year}
\usage{
real_income(base_year)
}
\arguments{
\item{base_year}{survey year. Find possible base years in the  \code{indinc_10} data set.}
}
\value{
A data frame with the inflated or deflated individual annual income for a selected year.
}
\description{
Provides the annual wage of an individual deflating or inflating annual income to an inputted base year (1989,1991,1993,1997,2000,2004,2006,2009,2011,2015)
}
\examples{
real_income(2015)
real_income(1989)
}
