% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_datasets.R
\docType{data}
\name{emw_12_clean}
\alias{emw_12_clean}
\title{Number of ever married women and number of children}
\format{
\subsection{\code{emw_12_clean}}{

A data frame with 26,616 rows and 60 columns:
\describe{
\item{IDind}{individual identification number}
\item{wave}{survey year}
\item{child_died}{Number of Children who died}
\item{child_births}{Total number of Children given births to}
...
}
#' @source \url{https://www.cpc.unc.edu/projects/china}
}
}
\usage{
emw_12_clean
}
\description{
Number of ever married women and number of children
}
\keyword{datasets}
