% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_datasets.R
\docType{data}
\name{indinc_10_clean}
\alias{indinc_10_clean}
\title{Annual Income from CHNS survey from 1989 to 2011}
\format{
A data frame with 62,029 rows and 32 columns.
\describe{
\item{IDind}{The IDind is a twelve-digit numeric variable that uniquely identified for all participants. Each participant will have the same ID in all datasets and in all survey years. }
\item{wave}{survey year}
\item{indwage}{nominal annual income}
}
}
\source{
\url{https://www.cpc.unc.edu/projects/china}
}
\usage{
indinc_10_clean
}
\description{
Annual Income from CHNS survey from 1989 to 2011
}
\keyword{datasets}
