% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_datasets.R
\docType{data}
\name{map_dat_clean}
\alias{map_dat_clean}
\title{Province-level average income data
Data set with annual average wage for the provinces and years included in the survey.}
\format{
A data frame that could map the average income by selected province and year.
\describe{
\item{province_en}{province names in English}
\item{wave}{survey year}
\item{wage_mean}{Annual average wage in unit of ¥1}
\item{name}{province names in Chinese}
\item{wage_mean_k}{Annual average income in unit of ¥1,000}
\item{wage_mean_cat}{Annual average income in unit of ¥1,000 divided into categories for mapping.}
}
}
\source{
\url{https://www.cpc.unc.edu/projects/china}
}
\usage{
map_dat_clean
}
\description{
Province-level average income data
Data set with annual average wage for the provinces and years included in the survey.
}
\keyword{datasets}
